refactor(media): extract FileVariantMixin from Movie and Episode
Deduplicate file variant helpers (primary_file, best_file,
available_resolutions, total_size, with_file, get_file_by_resolution)
into a shared mixin to prevent logic divergence.
